10/05/2024 - Character Stats Screen DONE - Navigation DONE - Character Equipment TODO

+++ BloodStats/app/src/main/java/com/sergimarrahyarenas/bloodstats/models/characterstatistics/Lifesteal.kt
@@ -2,6 +2,6 @@ package com.sergimarrahyarenas.bloodstats.models.characterstatistics
 
 data class Lifesteal(
     val rating: Int,
-    val rating_bonus: Int,
-    val value: Int
+    val rating_bonus: Double,
+    val value: Double
 )
