+++ BloodStats/app/src/main/java/com/sergimarrahyarenas/bloodstats/models/characterstatistics/Parry.kt
@@ -2,6 +2,6 @@ package com.sergimarrahyarenas.bloodstats.models.characterstatistics
 
 data class Parry(
     val rating: Int,
-    val rating_bonus: Int,
-    val value: Int
+    val rating_bonus: Double,
+    val value: Double
 )
